// src/enums/inventory.ts
export enum AllocatePatrs {
  ASSIGT_TO_WO = 'Assign to work order',
  REPLACE_ASSET = 'Replace Asset',
}

export enum ReplaceAssetAction {
  SELECT_FROM_LIST = 'Select From List',
  PULL_FROM_INVENTORY = 'Pull From Inventory',
  ADD_NEW_ASSET = 'Add New Asset',
}

export enum OldAssetActios {
  ARHIVE = 'Archive this asset',
  INVENTORY = 'Add to inventory',
}